+++ src/main/java/creatures/Dragon.java
@@ -2,7 +2,7 @@ package creatures;
 
 public class Dragon extends Creature {
 
-    public Dragon(String name, double defence, int attack) {
+    public Dragon(String name, int defence, int attack) {
         super(name, defence, attack);
     }
 }
