+++ src/main/java/creatures/Ogre.java
@@ -2,7 +2,7 @@ package creatures;
 
 public class Ogre extends Creature {
 
-    public Ogre(String name, double defence, int attack) {
+    public Ogre(String name, int defence, int attack) {
         super(name, defence, attack);
     }
 }
